+++ sbnet_tensorflow/sbnet_ops/reduce_mask.h
@@ -37,7 +37,7 @@ template <typename Device, typename T> struct ReduceMaskFunctor {
         int bCntW,   // Number of blocks, width.
         unsigned int numBins,
         unsigned int binSize,
-        long long* activeBlockIndices, // Indices of active blocks.
+        short* activeBlockIndices, // Triples of indices of active blocks.
         int* binCounts,
         bool avgPool
         );
